refactor: move computeWeekDaysFromFirstDay to post/utils
