+++ apps/web/src/shared/api/supabaseReads.ts
@@ -12,6 +12,7 @@ import { getSupabaseClient, isNetworkError, SupabaseNetworkError } from './supab
 import type { Board } from '@/board/model/Board';
 import type { Post } from '@/post/model/Post';
 import { PostVisibility } from '@/post/model/Post';
+import { computeWeekDaysFromFirstDay } from '@/post/utils/weekDays';
 import type { Comment } from '@/comment/model/Comment';
 import type { Reply } from '@/comment/model/Reply';
 import type { Reaction } from '@/comment/model/Reaction';
@@ -23,36 +24,6 @@ import { NotificationType } from '@/notification/model/Notification';
 function formatInFilter(values: string[]): string {
   const quoted = values.map((v) => `"${v.replace(/"/g, '""')}"`);
   return `(${quoted.join(',')})`;
-}
-
-/**
- * Compute the number of working days (Mon-Fri) from a board's first day to a post's creation date.
- *
- * Mirrors the Cloud Function `updatePostDaysFromFirstDay` with two differences:
- * - Uses postCreatedAt (fixed) instead of new Date() (current time).
- * - Computes KST weekday via modular arithmetic to avoid DST issues in non-KST browsers.
- *
- * Both dates are projected to KST (Asia/Seoul) calendar dates before counting.
- */
-export function computeWeekDaysFromFirstDay(boardFirstDay: string, postCreatedAt: string): number {
-  const kstStart = new Date(new Date(boardFirstDay).toLocaleString('en-US', { timeZone: 'Asia/Seoul' }));
-  const kstEnd = new Date(new Date(postCreatedAt).toLocaleString('en-US', { timeZone: 'Asia/Seoul' }));
-  kstStart.setHours(0, 0, 0, 0);
-  kstEnd.setHours(0, 0, 0, 0);
-
-  const msPerDay = 86400000;
-  const daysDiff = Math.ceil((kstEnd.getTime() - kstStart.getTime()) / msPerDay);
-
-  // Use modular arithmetic from the known KST start day to avoid creating
-  // intermediate Date objects that could be affected by local DST transitions.
-  const startDay = kstStart.getDay();
-  let workingDays = 0;
-  for (let i = 0; i < daysDiff; i++) {
-    const day = (startDay + i) % 7;
-    if (day !== 0 && day !== 6) workingDays++;
-  }
-
-  return workingDays;
 }
 
 // Supabase row types for query results
